+++ refena/lib/src/provider/types/notifier_provider.dart
@@ -22,6 +22,9 @@ class NotifierProvider<N extends BaseSyncNotifier<T>, T>
     return _build(ref, _builder);
   }
 
+  /// Overrides with a predefined notifier.
+  ///
+  /// {@category Initialization}
   ProviderOverride<N, T> overrideWithNotifier(N Function(Ref ref) builder) {
     return ProviderOverride(
       provider: this,